Added Gradle support and reorganized source tree for Gradle
