--- src/com/cleancoder/args/ArgumentMarshaler.java
@@ -1,7 +1,0 @@
-package com.cleancoder.args;
-
-import java.util.Iterator;
-
-public interface ArgumentMarshaler {
-  void set(Iterator<String> currentArgument) throws ArgsException;
-}
