+++ src/main/java/com/cleancoder/args/ArgumentMarshaler.java
@@ -1,0 +1,7 @@
+package com.cleancoder.args;
+
+import java.util.Iterator;
+
+public interface ArgumentMarshaler {
+  void set(Iterator<String> currentArgument) throws ArgsException;
+}
