+++ src/main/java/com/cleancoder/args/BooleanArgumentMarshaler.java
@@ -1,0 +1,18 @@
+package com.cleancoder.args;
+
+import java.util.Iterator;
+
+public class BooleanArgumentMarshaler implements ArgumentMarshaler {
+  private boolean booleanValue = false;
+
+  public void set(Iterator<String> currentArgument) throws ArgsException {
+    booleanValue = true;
+  }
+
+  public static boolean getValue(ArgumentMarshaler am) {
+    if (am != null && am instanceof BooleanArgumentMarshaler)
+      return ((BooleanArgumentMarshaler) am).booleanValue;
+    else
+      return false;
+  }
+}
