+++ src/main/java/smc/OptimizedStateMachine.java
