+++ src/main/java/smc/generators/CppCodeGenerator.java
