+++ src/main/java/smc/generators/JavaCodeGenerator.java
