+++ src/main/java/smc/generators/nestedSwitchCaseGenerator/NSCGenerator.java
