+++ src/main/java/smc/generators/nestedSwitchCaseGenerator/NSCNode.java
