+++ src/main/java/smc/generators/nestedSwitchCaseGenerator/NSCNodeVisitor.java
