+++ src/main/java/smc/implementers/CppNestedSwitchCaseImplementer.java
