+++ src/main/java/smc/lexer/Lexer.java
