+++ src/main/java/smc/lexer/TokenCollector.java
