+++ src/main/java/smc/optimizer/Optimizer.java
