+++ src/main/java/smc/parser/FsmSyntax.java
