+++ src/main/java/smc/semanticAnalyzer/SemanticStateMachine.java
