+++ src/test/java/smc/generators/nestedSwitchCaseGenerator/NSCGeneratorTest.java
