+++ src/test/java/smc/lexer/LexerTest.java
