+++ src/test/java/smc/optimizer/OptimizerTest.java
