+++ src/test/java/smc/semanticAnalyzer/SemanticAnalyzerTest.java
